Two new mappings and small changes
Added a mapping to to delete the rest of the paragraph with dp.

Added a mapping to quickly exit vim while multiple files are open.

Changed cases of <cr> to <CR> to conform to my standard.

Removed :call clearmatches() from <leader>n mapping as I think it might
have lead to the 80 character mark not being shown.