Save position of curser when formatting
It was annoying when formatting a paragraph of text to have to go to the
position you were previously in after formatting it as the macro did
not save the position. I now use vim marks to save the position and take
the curser back to where it started when formatting paragraphs.